Support handling all events in the client store, sent from the background store.
